+++ CHANGELOG.md
@@ -20,6 +20,7 @@ previous versions. You will need to upgrade your old database manually or using
 
 - New `data_type` attribute in HDF5 and TEXT spectrum files to identify
   the type of spectrum: `signal`, `noise`, `weight`, `synthetic`, `residual`.
+- Save spectral weights in residuals HDF5 file
 - Global inversion quality parameters in YAML, SQLite and HTML output:
   - Number of input stations
   - Number of input spectra